Add files via upload
Started by creating a basic class for Alien  Invasion that would create an instance of the game to actually run. This creates a blank window for our game assets to move on and look for inputs from our mouse and keyboard. Also included a little descriptor for the purpose of the actual project, as well as a reminder to myself to stay strong and keep up the coding. I every step, no matter how small, can be a step forward.