Added small code to change bg color
Added small bit of code to change the background color of the window to something more fitting for an alien invasion game. Also hope to add a class for a moving object to represent stars. Think it'd be pretty neat to have stars fly by while the ship travels in space. 
Added notes to help describe and understand how some of the functions and loop work to contribute to the whole project