Added ship class and drew ship on surface
Added a ship.py file to handle the ship class and it's settings/resources, also added code to alien_invasion to draw the ship on it's surface. assinged labels were kinda vague in the books example, so I made them more specific. referring to the actually object or element in the name. added some more notes and they are getting kinda out of hand, might start uploading a second edited file with no notes