Added code to move ship
Added a chunk of code to make the ship move to the right. used the event.type method to make the game listen for key strokes. currently have it set up so the ship will move right every time the right arrow key is pressed. will update later for continuous movement, also ship is just a big white block with a ship in the middle, will try to create a transparent layer so you don't see the square lol