Refactored check events and full screen
Refactored check events to to include a specific function for key down and key up events. also added ability to exit the game/kill the program using the esc key. Put the game in a full screen mode, which fixed some issues with the ship stuttering when it moved from left to right