Fire multiple bullets and they travel up screen
Made adjustments to the main alien invasion file so that it will fire multiple bullets and have them travel up screen. stores each instance of a bullet in a group and then runs the update method for each instance at once. Found out that the update method needs to be specifically defined as "update(argu)". Had it refactored as update_bullet and it did not run the update method for the Group